+++ checkstyle.xml
@@ -161,7 +161,7 @@
 <!--		<module name="CustomImportOrder">-->
 <!--			<property name="customImportOrderRules" value="STATIC###STANDARD_JAVA_PACKAGE###SPECIAL_IMPORTS###THIRD_PARTY_PACKAGE"/>-->
 <!--			<property name="standardPackageRegExp" value="^java\."/>-->
-		<!--			<property name="specialImportsRegExp" value="^jakarta\."/>-->
+<!--			<property name="specialImportsRegExp" value="^jakarta\."/>-->
 <!--			<property name="thirdPartyPackageRegExp" value="^org\."/>-->
 <!--			<property name="sortImportsInGroupAlphabetically" value="true"/>-->
 <!--			<property name="separateLineBetweenGroups" value="true"/>-->
