+++ gateway/src/main/java/ru/practicum/shareit/item/ItemClient.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.item;
+
+public class ItemClient {
+}
