+++ gateway/src/main/java/ru/practicum/shareit/request/ItemRequestClient.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.request;
+
+public class ItemRequestClient {
+}
