+++ server/src/main/java/ru/practicum/shareit/item/model/ItemDto.java
@@ -1,0 +1,43 @@
+package ru.practicum.shareit.item.model;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import ru.practicum.shareit.comment.model.CommentDto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ItemDto {
+    private long id;
+    @NotBlank(message = "Поле 'name' обязательно для заполнения")
+    private String name;
+    @NotBlank(message = "Поле 'description' обязательно для заполнения")
+    private String description;
+    @NotNull(message = "Поле 'available' обязательно для заполнения")
+    private Boolean available;
+    private LocalDateTime lastBooking;
+    private LocalDateTime nextBooking;
+    private List<CommentDto> comments;
+    private long requestId;
+
+    public ItemDto(long id, String name, String description, Boolean available) {
+        this.id = id;
+        this.name = name;
+        this.description = description;
+        this.available = available;
+    }
+
+    public ItemDto(long id, String name, String description, Boolean available, long requestId) {
+        this.id = id;
+        this.name = name;
+        this.description = description;
+        this.available = available;
+        this.requestId = requestId;
+    }
+}
