+++ server/src/main/java/ru/practicum/shareit/user/model/UserDto.java
@@ -1,0 +1,18 @@
+package ru.practicum.shareit.user.model;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserDto {
+
+    private long id;
+
+    private String name;
+
+    private String email;
+
+}
