--- src/main/java/ru/practicum/shareit/ShareItApp.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ShareItApp {
-    public static void main(String[] args) {
-        SpringApplication.run(ShareItApp.class, args);
-    }
-
-}
