--- src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -1,56 +1,0 @@
-package ru.practicum.shareit.booking;
-
-import jakarta.validation.Valid;
-import jakarta.validation.ValidationException;
-import jakarta.validation.constraints.NotNull;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.booking.model.BookingDto;
-import ru.practicum.shareit.booking.service.BookingService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping(path = "/bookings")
-@RequiredArgsConstructor
-public class BookingController {
-
-    @Autowired
-    private final BookingService bookingService;
-
-    @PostMapping
-    public BookingDto createBooking(@NotNull @RequestHeader("X-Sharer-User-Id") Long userId,
-                                    @Valid @RequestBody BookingDto bookingDto) {
-        if (bookingDto.getEnd().isBefore(bookingDto.getStart())) {
-            throw new ValidationException("Дата окончания бронирования не может быть раньше даты начала");
-        }
-
-        return bookingService.createBooking(bookingDto, userId);
-    }
-
-    @PatchMapping("/{bookingId}")
-    public BookingDto updateBookingStatus(@NotNull @RequestHeader("X-Sharer-User-Id") Long userId,
-                                          @PathVariable Long bookingId,
-                                          @RequestParam Boolean approved) {
-        return bookingService.updateBookingStatus(userId, bookingId, approved);
-    }
-
-    @GetMapping("/{bookingId}")
-    public BookingDto getBooking(@NotNull @RequestHeader("X-Sharer-User-Id") Long userId,
-                                 @PathVariable Long bookingId) {
-        return bookingService.getBookingById(userId, bookingId);
-    }
-
-    @GetMapping
-    public List<BookingDto> getUserBookings(@NotNull @RequestHeader("X-Sharer-User-Id") Long userId,
-                                            @RequestParam(defaultValue = "ALL") String state) {
-        return bookingService.getUserBookings(userId, state);
-    }
-
-    @GetMapping("/owner")
-    public List<BookingDto> getOwnerBookings(@NotNull @RequestHeader("X-Sharer-User-Id") Long userId,
-                                             @RequestParam(defaultValue = "ALL") String state) {
-        return bookingService.getOwnerBookings(userId, state);
-    }
-}
