--- src/main/java/ru/practicum/shareit/booking/model/BookingMapper.java
@@ -1,31 +1,0 @@
-package ru.practicum.shareit.booking.model;
-
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.item.model.ItemMapper;
-import ru.practicum.shareit.user.model.User;
-import ru.practicum.shareit.user.model.UserMapper;
-
-public class BookingMapper {
-
-    public static Booking toBooking(BookingDto dto, User booker, Item item, BookingStatus status) {
-        Booking booking = new Booking();
-        booking.setId(dto.getId());
-        booking.setStart(dto.getStart());
-        booking.setEnd(dto.getEnd());
-        booking.setBooker(booker);
-        booking.setItem(item);
-        booking.setStatus(status);
-        return booking;
-    }
-
-    public static BookingDto toBookingDto(Booking booking) {
-        BookingDto dto = new BookingDto();
-        dto.setId(booking.getId());
-        dto.setStart(booking.getStart());
-        dto.setEnd(booking.getEnd());
-        dto.setStatus(booking.getStatus());
-        dto.setBooker(UserMapper.toUserDto(booking.getBooker()));
-        dto.setItem(ItemMapper.toItemDto(booking.getItem()));
-        return dto;
-    }
-}
