--- src/main/java/ru/practicum/shareit/booking/service/BookingService.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.booking.service;
-
-import ru.practicum.shareit.booking.model.BookingDto;
-
-import java.util.List;
-
-public interface BookingService {
-    BookingDto createBooking(BookingDto bookingDto, Long userId);
-
-    BookingDto updateBookingStatus(Long userId, Long bookingId, Boolean approved);
-
-    BookingDto getBookingById(Long userId, Long bookingId);
-
-    List<BookingDto> getUserBookings(Long userId, String state);
-
-    List<BookingDto> getOwnerBookings(Long ownerId, String state);
-}
