--- src/main/java/ru/practicum/shareit/comment/model/Comment.java
@@ -1,37 +1,0 @@
-package ru.practicum.shareit.comment.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.model.User;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "comments")
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class Comment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "text")
-    private String text;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id")
-    private Item item;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "author_id")
-    private User author;
-
-    @Column(name = "created")
-    private LocalDateTime created;
-}
