--- src/main/java/ru/practicum/shareit/comment/model/CommentMapper.java
@@ -1,16 +1,0 @@
-package ru.practicum.shareit.comment.model;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class CommentMapper {
-    public static CommentDto toCommentDto(Comment comment) {
-        return new CommentDto(
-                comment.getId(),
-                comment.getText(),
-                comment.getAuthor().getName(),
-                comment.getCreated()
-        );
-    }
-}
