--- src/main/java/ru/practicum/shareit/comment/repository/CommentRepository.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit.comment.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.comment.model.Comment;
-
-import java.util.List;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    List<Comment> findByItemIdOrderByCreatedDesc(Long itemId);
-}
-
