--- src/main/java/ru/practicum/shareit/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(String message) {
-        super(message);
-    }
-
-}
