--- src/main/java/ru/practicum/shareit/exception/ConflictException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class ConflictException extends RuntimeException {
-    public ConflictException(String message) {
-        super(message);
-    }
-}
