--- src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -1,65 +1,0 @@
-package ru.practicum.shareit.item;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.comment.model.CommentDto;
-import ru.practicum.shareit.item.model.ItemDto;
-import ru.practicum.shareit.item.service.ItemService;
-
-import java.util.Collection;
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/items")
-public class ItemController {
-
-    @Autowired
-    private final ItemService itemService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public ItemDto createItem(@Valid @RequestBody ItemDto itemDto, @NotNull @RequestHeader("X-Sharer-User-Id") long userId) {
-        return itemService.createItem(itemDto, userId);
-    }
-
-    @PatchMapping("/{itemId}")
-    @ResponseStatus(HttpStatus.OK)
-    public ItemDto updateItem(
-            @RequestBody ItemDto itemDto,
-            @PathVariable("itemId") long itemId,
-            @NotNull @RequestHeader("X-Sharer-User-Id") long userId) {
-        return itemService.updateItem(itemDto, itemId, userId);
-    }
-
-    @GetMapping("/{itemId}")
-    @ResponseStatus(HttpStatus.OK)
-    public ItemDto getItemById(@PathVariable("itemId") long itemId) {
-        return itemService.getItemById(itemId);
-    }
-
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public List<ItemDto> getItems(@NotNull @RequestHeader("X-Sharer-User-Id") long userId) {
-        return itemService.getItems(userId);
-    }
-
-    @GetMapping("/search")
-    @ResponseStatus(HttpStatus.OK)
-    public Collection<ItemDto> searchItems(@RequestParam("text") String text) {
-        return itemService.searchItems(text);
-    }
-
-    @PostMapping("/{itemId}/comment")
-    @ResponseStatus(HttpStatus.CREATED)
-    public CommentDto addComment(
-            @NotNull @RequestHeader("X-Sharer-User-Id") long userId,
-            @PathVariable long itemId,
-            @Valid @RequestBody CommentDto commentDto) {
-        return itemService.addComment(userId, itemId, commentDto);
-    }
-}
