--- src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,43 +1,0 @@
-package ru.practicum.shareit.item.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-import ru.practicum.shareit.request.model.ItemRequest;
-
-@Entity
-@Table(name = "items")
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class Item {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "available")
-    private boolean available;
-
-    @Column(name = "owner")
-    private long owner;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "request")
-    private ItemRequest request;
-
-    public Item(long id, String name, String description, boolean available, long owner) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.available = available;
-        this.owner = owner;
-    }
-}
