--- src/main/java/ru/practicum/shareit/item/model/ItemMapper.java
@@ -1,38 +1,0 @@
-package ru.practicum.shareit.item.model;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class ItemMapper {
-    public static Item toItem(ItemDto itemDto, long id, long userId) {
-        return new Item(
-                id,
-                itemDto.getName(),
-                itemDto.getDescription(),
-                itemDto.getAvailable(),
-                userId
-        );
-    }
-
-    public static ItemDto toItemDto(Item item) {
-        return new ItemDto(
-                item.getId(),
-                item.getName(),
-                item.getDescription(),
-                item.isAvailable()
-        );
-    }
-
-    public static void updateItemFields(Item item, ItemDto request) {
-        if (request.getName() != null && !request.getName().isBlank()) {
-            item.setName(request.getName());
-        }
-        if (request.getDescription() != null && !request.getDescription().isBlank()) {
-            item.setDescription(request.getDescription());
-        }
-        if (request.getAvailable() != null) {
-            item.setAvailable(request.getAvailable());
-        }
-    }
-}
