--- src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,29 +1,0 @@
-package ru.practicum.shareit.request.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "item_requests")
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class ItemRequest {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "requestor")
-    private Long requestor;
-
-    @Column(name = "created")
-    private LocalDateTime created;
-}
