--- src/main/java/ru/practicum/shareit/user/UserController.java
@@ -1,50 +1,0 @@
-package ru.practicum.shareit.user;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.user.model.UserDto;
-import ru.practicum.shareit.user.service.UserService;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping(path = "/users")
-public class UserController {
-
-    @Autowired
-    private final UserService userService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDto createUser(@Valid @RequestBody UserDto userDto) {
-        return userService.createUser(userDto);
-    }
-
-    @PatchMapping("/{userId}")
-    @ResponseStatus(HttpStatus.OK)
-    public UserDto updateUser(@PathVariable("userId") long userId, @RequestBody UserDto userDto) {
-        return userService.updateUser(userId, userDto);
-    }
-
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public List<UserDto> getUsers() {
-        return userService.getUsers();
-    }
-
-    @GetMapping("/{userId}")
-    @ResponseStatus(HttpStatus.OK)
-    public UserDto getUserById(@PathVariable("userId") long userId) {
-        return userService.getUserById(userId);
-    }
-
-    @DeleteMapping("/{userId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable("userId") long userId) {
-        userService.deleteUser(userId);
-    }
-}
