--- src/main/java/ru/practicum/shareit/user/model/UserMapper.java
@@ -1,32 +1,0 @@
-package ru.practicum.shareit.user.model;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class UserMapper {
-    public static User toUser(UserDto userDto, long id) {
-        return new User(
-                id,
-                userDto.getName(),
-                userDto.getEmail()
-        );
-    }
-
-    public static UserDto toUserDto(User user) {
-        return new UserDto(
-                user.getId(),
-                user.getName(),
-                user.getEmail()
-        );
-    }
-
-    public static void updateUserFields(User user, UserDto request) {
-        if (request.getName() != null && !request.getName().isBlank()) {
-            user.setName(request.getName());
-        }
-        if (request.getEmail() != null && !request.getEmail().isBlank()) {
-            user.setEmail(request.getEmail());
-        }
-    }
-}
