--- src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package ru.practicum.shareit.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.user.model.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
-}
