--- src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -1,18 +1,0 @@
-package ru.practicum.shareit.user.service;
-
-
-import ru.practicum.shareit.user.model.UserDto;
-
-import java.util.List;
-
-public interface UserService {
-    UserDto createUser(UserDto userDto);
-
-    UserDto updateUser(long userId, UserDto userDto);
-
-    UserDto getUserById(long userId);
-
-    List<UserDto> getUsers();
-
-    void deleteUser(long userId);
-}
