--- src/main/resources/application.properties
@@ -1,23 +1,0 @@
-spring.jpa.hibernate.ddl-auto=create
-spring.jpa.properties.hibernate.format_sql=true
-spring.sql.init.mode=always
-
-logging.level.org.springframework.transaction.interceptor=INFO
-logging.level.org.springframework.orm.jpa.JpaTransactionManager=INFO
-logging.level.org.hibernate.SQL=INFO
-logging.level.org.hibernate.type.descriptor.sql=INFO
-logging.level.org.springframework.orm.jpa=INFO
-logging.level.org.springframework.transaction=INFO
-
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/shareit
-spring.datasource.username=postgres
-spring.datasource.password=12345
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-#---
-spring.config.activate.on-profile=ci,test
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.url=jdbc:h2:mem:shareit
-spring.datasource.username=postgres
-spring.datasource.password=12345
