--- src/main/resources/schema.sql
@@ -1,51 +1,0 @@
-CREATE TABLE IF NOT EXISTS users (
-  id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-  name VARCHAR(255) NOT NULL,
-  email VARCHAR(512) NOT NULL,
-  CONSTRAINT pk_users PRIMARY KEY (id),
-  CONSTRAINT UQ_USER_EMAIL UNIQUE (email)
-);
-
-CREATE TABLE IF NOT EXISTS item_requests (
-    id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    description TEXT NOT NULL,
-    requestor BIGINT NOT NULL,
-    created TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    CONSTRAINT pk_item_requests PRIMARY KEY (id),
-    CONSTRAINT fk_requestor FOREIGN KEY (requestor) REFERENCES users (id) ON DELETE CASCADE
-);
-
-CREATE TABLE IF NOT EXISTS items (
-  id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-  name VARCHAR(255) NOT NULL,
-  description TEXT,
-  available BOOLEAN NOT NULL,
-  owner BIGINT NOT NULL,
-  request BIGINT,
-  CONSTRAINT pk_items PRIMARY KEY (id),
-  CONSTRAINT fk_owner FOREIGN KEY (owner) REFERENCES users (id) ON DELETE CASCADE,
-  CONSTRAINT fk_request FOREIGN KEY (request) REFERENCES item_requests (id) ON DELETE CASCADE
-);
-
-CREATE TABLE IF NOT EXISTS bookings (
-    id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    start_date TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    end_date TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    item_id BIGINT NOT NULL,
-    booker_id BIGINT NOT NULL,
-    status VARCHAR(50) NOT NULL,
-    CONSTRAINT pk_bookings PRIMARY KEY (id),
-    CONSTRAINT fk_item_id FOREIGN KEY (item_id) REFERENCES items (id) ON DELETE CASCADE,
-    CONSTRAINT fk_booker_id FOREIGN KEY (booker_id) REFERENCES users (id) ON DELETE CASCADE
-);
-
-CREATE TABLE IF NOT EXISTS comments (
-    id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    text TEXT NOT NULL,
-    item_id BIGINT NOT NULL,
-    author_id BIGINT NOT NULL,
-    created TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-    CONSTRAINT pk_comments PRIMARY KEY (id),
-    CONSTRAINT fk_comments_item_id FOREIGN KEY (item_id) REFERENCES items (id) ON DELETE CASCADE,
-    CONSTRAINT fk_comments_author_id FOREIGN KEY (author_id) REFERENCES users (id) ON DELETE CASCADE
-);
